Move All Issues tab contents to their own file
