+++ client/src/app/pages/issues/single-app-issues-tab.tsx
@@ -1,0 +1,7 @@
+import * as React from "react";
+
+export interface ISingleAppIssuesTabProps {}
+
+export const SingleAppIssuesTab: React.FC<ISingleAppIssuesTabProps> = () => (
+  <>TODO</>
+);
